requirements.txt
PyYAML
nibabel
pyvista
scikit-image
SimpleITK
numpy
totalsegmentator
xgboost